Feedback: removed comments, used re-usable function for token
