+++ src/controllers/travelRequest.controller.js
@@ -1,5 +1,5 @@
-import db from '../models'
 import { getDataFromToken } from '../helper/tokenToData';
+import { createTravelRequest } from '../services/createTravelRequest';
 
 const travelRequest = async (req, res, next) => {
     const decoded = await getDataFromToken(req, res, next)
@@ -12,39 +12,12 @@ const travelRequest = async (req, res, next) => {
         const request = {
             managerId:decoded.manager_id,
             userId:decoded.id,
-            // status:req.body.status,
             createdAt:new Date(),
             updatedAt: new Date(),
         }
-        /*const t = db.sequelize.transaction()*/
         try{
-            db.TravelRequest.create(request)
-            .then(tRequestData => {
-                try{
-                    var counter = req.body.trip.length
-                    for(var record of req.body.trip){
-                        counter -=1
-                        record.travelId = tRequestData.travelId
-                        if(counter == 0){
-                            db.Trip.bulkCreate(req.body.trip, {hooks:true})
-                            .then((tripData) => { 
-                                /*awit.commit();*/ 
-                                var allData = Object.assign({}, tRequestData.get({ plain: true }), {tripData})
-                                res.json({message: "Trip request sent successfully", data: allData})})
-                            .catch((err) => {/*await t.rollback();res.json(err);*/  next(err);  console.log(err.message);})
-                        }
-                    }
-                }catch(err){
-                    // res.status(400).json(err)
-                    next(err)
-                }
-            })
-            .catch(err => {
-                // res.json(err)
-                next(err)
-            })
+            createTravelRequest(req, res, request, next)
         }catch(err){
-            // res.status(400).json(rr.message)
             next(err)
         }
     }else if(decoded.managerId == undefined){
