+++ src/models/travelrequest.js
@@ -1,24 +1,6 @@
 'use strict';
 
-// const {
-//   Model
-// } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
-//   class TravelRequest extends Model {
-//     /**
-//      * Helper method for defining associations.
-//      * This method is not a part of Sequelize lifecycle.
-//      * The `models/index` file will call this method automatically.
-//      */
-//     static associate(models) {
-//       // define association here
-      
-//       // models.Trip.hasOne(TravelRequest, {foreignKey:"travelId"})
-//       // TravelRequest.belongsToMany(models.Trip, {foreignKey:"travelId"})
-//       // TravelRequest.sync({alter:true})
-//     }
-//   };
-//   TravelRequest.init({
   const TravelRequest = sequelize.define('TravelRequest',{
     travelId: {
       allowNull: false,
@@ -34,10 +16,6 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
       type:DataTypes.UUID
     },
-    // tripId: {
-    //   allowNull: false,
-    //   type:DataTypes.INTEGER
-    // },
     status: {
       allowNull: false,
       defaultValue: "pending",
