+++ src/models/trip.js
@@ -1,21 +1,5 @@
 'use strict';
-// const {Model} = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
-//   class Trip extends Model {
-//     /**
-//      * Helper method for defining associations.
-//      * This method is not a part of Sequelize lifecycle.
-//      * The `models/index` file will call this method automatically.
-//      */
-//     static associate(models) {
-//       // define association here
-//       // models.TravelRequest.hasMany(Trip, {foreignKey:"travelId", targetKey:"travelId"})
-//       // Trip.belongsTo(models.TravelRequest, {foreignKey:"travelId", targetKey:"travelId"})
-//       // Trip.sync({alter:true})
-
-//     }
-//   };
-  // Trip.init({
     const Trip = sequelize.define('Trip', {
     tripId: {
       allowNull: false,
@@ -46,10 +30,6 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.UUID,
       allowNull: false
     },
-    // travelId: {
-    //   type: DataTypes.UUID,
-    //   allowNull: false
-    // },
     createdAt: {
       allowNull: false,
       type: DataTypes.DATE
@@ -67,7 +47,6 @@ module.exports = (sequelize, DataTypes) => {
   Trip.associate = (models) => {
     Trip.belongsTo(models.TravelRequest, {
       foreignKey:'travelId',
-      // as: 'travelRequest'
     })
   }
   return Trip;
